requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
torch
torchvision
pillow
python-multipart